python/extra/requirements.txt
# API Clients
requests>=2.28.1
supabase>=1.0.3
python-dotenv>=0.21.0

# Rate Limiting and Caching
tenacity>=8.1.0
requests-ratelimiter>=0.4.0
diskcache>=5.4.0

# Utilities
argparse>=1.4.0 